sptrans: treat hex within CSUB as identifiers (#33)
The values within a CSUB are not correctly formatted BASIC hex numbers,
i.e. they don't being &h or &H.

Workaround this by flagging all numbers within a CSUB definition as
identifiers.

It's a bit of a HACK because the lexer is now doing a bit of syntax
analysis - and not in a sophisticated way as it is just flipping a
flag each time it encounters the CSUB keyword.

Closes-Bug: #33