splib: workaround controller polling issues on MMB4L
Games using ctrl.poll_multiple%() to determine which controller was
in use were very unresponsive to the initial button press when running
under MMB4L's PicoMiteVGA simulation on the R36S console. I suspect
this was due to the time taken to open/close a gamepad (via SDL) being
so much less responsive than the direct GPIO mechanism on the real
hardware.